TL;DR/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
transformers==4.34.1
